Refactor route 'api/todolists/' into 'api/todo-list/'
