+++ project/urls.py
@@ -17,7 +17,7 @@ from django.contrib import admin
 from django.views.generic import TemplateView
 from django.urls import re_path, path, include
 
-from todolist import urls as todolist_enpoints
+from todo_list import urls as todo_list_enpoints
 
 frontend_urls = [
     path("", include("pwa.urls")),
@@ -33,6 +33,6 @@ urlpatterns = [
     ),
     path("api/rest-auth/registration/", include("rest_auth.registration.urls")),
     path("api/rest-auth/", include("rest_auth.urls")),
-    path("api/todolists/", include(todolist_enpoints)),
+    path("api/todo-list/", include(todo_list_enpoints)),
     path("admin/", admin.site.urls),
 ] + frontend_urls
