+++ todo_list/apps.py
@@ -2,4 +2,4 @@ from django.apps import AppConfig
 
 
 class TodoListConfig(AppConfig):
-    name = "todolist"
+    name = "todo_list"
